[Core/VMAP] Use bounding geometry in vmaps.
This patch fixes 3 issues with the vmap extractor.

1. Incorrectly uses full model geometry instead of bounding geometry.
2. Incorrectly converts vertex coordinates.
3. Forgets to convert coordinates on triangle indices.

The effects of [1] are that LOS issues appear anywhere geometry is used
for decoration i.e. tree leaves, spider webs.

[2] Causes models to be flipped. This isn't very noticable on most trees
as flipping a cylinder results in the same cylinder, but it's very
noticable on any non-symetical geometry (which, even includes trees,
it's just harder to notice).

[3] Didn't seem to be a problem when the coordinates were incorrectly
converted, but when applying the correct conversion caused some
triangles to be flipped (i.e. facing inwards).

You will need to re-extract vmaps and re-generate mmaps for these
changes to take effect.

Signed-off-by: Daniel S. Reichenbach <daniel@kogitoapp.com>