[Fix] lift check for facing targets out of range
Whether the caster is facing the target or not has nothing to do with
whether the target is in range. CheckRange returned early for spells
with SPELL_RANGE_IDX_COMBAT, skipping the facing-target check.

Also remove code that allows spells triggered by auras to be cast
regardless of range, reverting part of [z0555].

Signed-off-by: Daniel S. Reichenbach <daniel@kogitoapp.com>