[Fix] Fix creatures running in the wrong direction
Two issues can cause creatures to appear to run in the opposite
direction of where they are really heading.

If a SMSG_MONSTER_MOVE with MonsterMoveFacingAngle is sent immediately
before a MonsterMoveNormal, the creature will seem to run in the
direction of MonsterMoveFacingAngle, rather than the proper direction.
This occured because MonsterMoveFacingAngle was used instead of
MonsterMoveStop in Unit::StopMoving.

A MSG_MOVE_SET_X_SPEED can also cause the issue. Since the speed update
is handled in SMSG_UPDATE_OBJECT these packets do not need to be sent.

Signed-off-by: Daniel S. Reichenbach <daniel@kogitoapp.com>