[Core/Units] Merge InterruptMoving into StopMoving
Because StopMoving no longer changes the unit's spline, units
continued to move after it was called. This is not desirable behaviour
under any condition so the code that stopped the unit from moving
server-side in InterruptMoving is now part of StopMoving.

InterruptMoving has been removed since all of its functionality is now
part of StopMoving. StopMoving should be used wherever InterruptMoving
was used.

Signed-off-by: Daniel S. Reichenbach <daniel@kogitoapp.com>