[Fix] aggro caused stunned creatures facing wrong
If a creature is aggroed with a stun, the creature will begin chasing
the player and get stunned in the same update, in that order. So when
the stun causes the creature to stop moving, it will have its
orientation turned toward the player.

To fix it, creatures shouldn't get a new orientation in
Unit::StopMoving when they are currently stunned.

Signed-off-by: Daniel S. Reichenbach <daniel@kogitoapp.com>